Added login/signup support
This is a large change. In this change I've added login and signup
support for different roles. There are primarily 3 roles: 1. user, 2.
admin, and 3. moderator.
I've also defined a user-model to store the information related to the
user. Each user will have the following data: 1. username, 2. email, 3.
password, and roles.
This commit also adds the 'middleware' directory in the root folder that
contains middleware files associated with authentication and signup.
Next, I've also added some controllers to render the content based on
different roles. And the controllers directory now also contains code
for authorizing signups, and signin.
Not only this, now we have auth-routes.js and user-routes.js files in
the routes directory, these files are not complete yet. Additional work
is still needed to fully implement this functionality.